bpf/headers: widen session flags to __u16 for NPTv6 bit
SESS_FLAG_NPTV6 is defined as (1 << 8) = 256, but the session_value
flags field it belongs to was __u8 (0-255). Bits 0-7 are fully consumed
by SNAT..NAT64, so bit 8 overflows the byte: any attempt to set the
NPTv6 flag truncates to 0 (colliding with "no flags"). The defect is
latent today -- no runtime code writes SESS_FLAG_NPTV6, and the live
NPTv6 path (userspace-dp/src/nptv6.rs) is stateless RFC 6296 prefix
translation that never touches a session flag -- but the constant is a
footgun the moment session-level NPTv6 flagging is added.

There is no free bit <= 7 to relocate into, so widen the flags field
from __u8 to __u16 across every mirror of the shared conntrack struct,
keeping the C/Rust/Go byte layout in lockstep:

  - C  struct session_value{,_v6}.flags __u8 -> __u16
  - Rust BpfSessionValueV4/V6.flags u8 -> u16 (+ SESS_FLAG_* consts)
  - Go  SessionValue{,V6}.Flags and the on-map bpfSessionValue{,V6}.Flags
        uint8 -> uint16 (+ SessFlagNAT64/SessFlagNPTV6 consts)

The compiler pads the leading state/flags/tcp_state/is_reverse block
from 8 to 16 bytes, so the on-map conntrack ABI grows 128 -> 136 (v4)
and 176 -> 184 (v6). All three size asserts are bumped in lockstep
(bpf_map_tests.rs, bpf_session_value_test.go). The map value_size is
derived at runtime from sizeof(bpfSessionValue) (loader_userspace_shim
.go), so the registration follows automatically; the map is node-local,
so growing it is not a cross-node concern.

The retained AF_XDP shim ABI is unaffected -- its userspace_sessions map
is HashMap<UserspaceSessionKey, u8> and it does not reference
session_value -- so `make generate` is not required.

HA session-sync wire is kept byte-identical: the Flags field stays one
byte at its fixed offset (low 8 bits). Every flag a session sets today
is in bits 0-7, and bits >= 8 are never set at runtime, so the low-byte
encoding is loss-free and needs no wire version bump. A comment records
that a future NPTv6 session flag must be carried as a length-gated
trailing field (like Generation) rather than widening the fixed byte.

Validation:
  - New counter-factual RED-on-revert tests
    nptv6_session_flag_survives_roundtrip_v4/v6 write bit 8 into the
    flags field, round-trip through the on-map bytes, and assert it
    survives. Verified firsthand that reverting the production width to
    __u8 makes both FAIL with a clean assertion (not a build break);
    the struct-size asserts flip to 136/184 too.
  - cargo build --release + cargo test nptv6 green.
  - go build ./... and go test ./pkg/dataplane/... ./pkg/cluster/...
    ./pkg/api/... ./pkg/grpcapi/... ./pkg/daemon/... green;
    go vet ./... clean.

Closes #5460.

+++ pkg/cluster/sync_protocol.go
@@ -112,7 +112,13 @@ func encodeSessionV4Payload(key dataplane.SessionKey, val dataplane.SessionValue
 	off += 4
 	buf[off] = val.State
 	off++
-	buf[off] = val.Flags
+	// #5460: SessionValue.Flags is uint16, but the HA sync wire carries only the
+	// low byte here. All SESS_FLAG_* bits currently set on a session are 0-7
+	// (SNAT..NAT64); bit 8 (SessFlagNPTV6) is never set today, so truncating to a
+	// byte is loss-free. If session-level NPTv6 flagging is ever added, carry the
+	// high byte as a trailing length-gated field (like Generation) rather than
+	// widening this fixed-offset byte in place — that would be a wire flag-day.
+	buf[off] = byte(val.Flags)
 	off++
 	buf[off] = val.TCPState
 	off++
@@ -207,7 +213,13 @@ func encodeSessionV6Payload(key dataplane.SessionKeyV6, val dataplane.SessionVal
 	off += 4
 	buf[off] = val.State
 	off++
-	buf[off] = val.Flags
+	// #5460: SessionValue.Flags is uint16, but the HA sync wire carries only the
+	// low byte here. All SESS_FLAG_* bits currently set on a session are 0-7
+	// (SNAT..NAT64); bit 8 (SessFlagNPTV6) is never set today, so truncating to a
+	// byte is loss-free. If session-level NPTv6 flagging is ever added, carry the
+	// high byte as a trailing length-gated field (like Generation) rather than
+	// widening this fixed-offset byte in place — that would be a wire flag-day.
+	buf[off] = byte(val.Flags)
 	off++
 	buf[off] = val.TCPState
 	off++
@@ -356,7 +368,9 @@ func decodeSessionV4Payload(payload []byte) (dataplane.SessionKey, dataplane.Ses
 	}
 	val.State = payload[off]
 	off++
-	val.Flags = payload[off]
+	// #5460: low byte only — see the encoder. Flags is uint16; the high byte
+	// (bits >=8, SessFlagNPTV6) is not carried on the wire and defaults to 0.
+	val.Flags = uint16(payload[off])
 	off++
 	val.TCPState = payload[off]
 	off++
@@ -470,7 +484,9 @@ func decodeSessionV6Payload(payload []byte) (dataplane.SessionKeyV6, dataplane.S
 	}
 	val.State = payload[off]
 	off++
-	val.Flags = payload[off]
+	// #5460: low byte only — see the encoder. Flags is uint16; the high byte
+	// (bits >=8, SessFlagNPTV6) is not carried on the wire and defaults to 0.
+	val.Flags = uint16(payload[off])
 	off++
 	val.TCPState = payload[off]
 	off++
@@ -713,7 +729,7 @@ func stripFullSetSeq(payload []byte) (base []byte, incarnation, seq uint64) {
 		bytes.Equal(payload[n-fullSetSeqTrailerLen:n-16], fullSetSeqMagic[:]) {
 		incarnation = binary.LittleEndian.Uint64(payload[n-16 : n-8])
 		seq = binary.LittleEndian.Uint64(payload[n-8:])
-		return payload[: n-fullSetSeqTrailerLen], incarnation, seq
+		return payload[:n-fullSetSeqTrailerLen], incarnation, seq
 	}
 	return payload, 0, 0
 }
